[F] refactor MyApi list
