+++ packages/frontend/src/types/Api.ts
@@ -4,3 +4,11 @@ export type ApiDef = {
   createdAt: string;
   updatedAt: string;
 };
+
+export type ApiListItem = ApiDef & { status: string };
+
+export type ApiList = {
+  list: ApiListItem[];
+  onDelete: (index: number) => any;
+  onUpdate: (index: number) => any;
+};
