--- packages/frontend/src/view/MyAPI/List/types.ts
@@ -1,5 +1,0 @@
-export type ApiDefsListFC = {
-  list: string[][];
-  onDelete: (index: number) => any;
-  onUpdate: (index: number) => any;
-};
