+++ packages/frontend/src/view/MyAPI/MyApi/index.ts
@@ -1,0 +1,1 @@
+export { MyAPI } from './MyAPI';
